create PlotArray helper class
