+++ python/src/scipp/plot/plot.py
@@ -6,50 +6,9 @@ from .._scipp import core as sc
 from .. import _utils as su
 from ..compat.dict import from_dict
 from .dispatch import dispatch
+from .helpers import Plot
 from .tools import make_fake_coord, get_line_param
 import numpy as np
-
-
-class Plot(dict):
-    """
-    The Plot object is used as output for the plot command.
-    It is a small wrapper around python dict, with an `_ipython_display_`
-    representation.
-    The dict will contain one entry for each entry in the input supplied to
-    the plot function.
-    More functionalities can be added in the future.
-    """
-    def __init__(self, *arg, **kw):
-        super(Plot, self).__init__(*arg, **kw)
-
-    def _ipython_display_(self):
-        """
-        IPython display representation for Jupyter notebooks.
-        """
-        return self._to_widget()._ipython_display_()
-
-    def _to_widget(self):
-        """
-        Return plot contents into a single VBocx container
-        """
-        import ipywidgets as ipw
-        contents = []
-        for item in self.values():
-            if item is not None:
-                contents.append(item._to_widget())
-        return ipw.VBox(contents)
-
-    def show(self):
-        for item in self.values():
-            item.show()
-
-    def as_static(self, *args, **kwargs):
-        """
-        Convert all the contents of the dict to static plots, releasing the
-        memory held by the plot (which makes a full copy of the input data).
-        """
-        for key, item in self.items():
-            self[key] = item.as_static(*args, **kwargs)
 
 
 def _variable_to_data_array(variable):
